+++ History.md
@@ -4,7 +4,9 @@ Dalli Changelog
 HEAD
 =======
 
- - Add support for >1MB values with :value_max_bytes option (r-stu31)
+ - Add support for username/password as options, to allow multiple bucket access
+   from the same Ruby process, GH-52
+ - Add support for >1MB values with :value_max_bytes option, GH-54 (r-stu31)
  - Add support for default TTL, :expires_in, in Rails 2.3. (Steven Novotny)
    config.cache_store = :dalli_store, 'localhost:11211', {:expires_in => 4.hours}